+++ src/apps/kwenta/kwenta.definition.ts
@@ -20,6 +20,16 @@ export const KWENTA_DEFINITION = appDefinition({
       type: GroupType.POSITION,
       label: 'Cross',
     },
+    staking: {
+      id: 'staking',
+      type: GroupType.POSITION,
+      label: 'Staking',
+    },
+    lpStaking: {
+      id: 'lp-staking',
+      type: GroupType.POSITION,
+      label: 'LP Staking',
+    },
   },
   tags: [AppTag.DERIVATIVES],
   keywords: [],
@@ -34,6 +44,10 @@ export const KWENTA_DEFINITION = appDefinition({
   },
 
   primaryColor: '#fff',
+  token: {
+    address: '0x920cf626a271321c151d027030d5d08af699456b',
+    network: Network.OPTIMISM_MAINNET,
+  },
 });
 
 @Register.AppDefinition(KWENTA_DEFINITION.id)