+++ src/apps/kwenta/kwenta.module.ts
@@ -5,6 +5,8 @@ import { KwentaContractFactory } from './contracts';
 import { KwentaAppDefinition, KWENTA_DEFINITION } from './kwenta.definition';
 import { OptimismKwentaCrossContractPositionFetcher } from './optimism/kwenta.cross.contract-position-fetcher';
 import { OptimismKwentaIsolatedContractPositionFetcher } from './optimism/kwenta.isolated.contract-position-fetcher';
+import { OptimismKwentaLpStakingContractPositionFetcher } from './optimism/kwenta.lp-staking.contract-position-fetcher';
+import { OptimismKwentaStakingContractPositionFetcher } from './optimism/kwenta.staking.contract-position-fetcher';
 
 @Register.AppModule({
   appId: KWENTA_DEFINITION.id,
@@ -13,6 +15,8 @@ import { OptimismKwentaIsolatedContractPositionFetcher } from './optimism/kwenta
     KwentaContractFactory,
     OptimismKwentaIsolatedContractPositionFetcher,
     OptimismKwentaCrossContractPositionFetcher,
+    OptimismKwentaStakingContractPositionFetcher,
+    OptimismKwentaLpStakingContractPositionFetcher,
   ],
 })
 export class KwentaAppModule extends AbstractApp() {}